// frontend/src/app/model/appointment.ts
export class Appointment{
    patientUsername: String;
    patientFirstname: String;
    patientLastname: String;
    doctorUsername: String;
    doctorFirstname: String;
    doctorLastname: String;
    doctorBranch: String;
    selectedDate: String;
    selectedTime: String;
    examTitle: String;
    examDuration: String;
    examPrice: String;
}